Put function which might be useful across many files in a separate file and folder src/api/general.js
